+++ CHANGELOG.rst
@@ -1,8 +1,32 @@
 Changelog
 =========
 
+0.11.4 (2023-12-08)
+-------------------
+
+- Modified `PackageURL.from_string` to properly handle golang purls.
+
+0.11.3 (2023-12-08)
+--------------------
+
+- Add support for GitLab "/archive/" URLs in `url2purl`.
+  https://github.com/package-url/packageurl-python/issues/133
+
+0.11.2 (2022-07-25)
+--------------------
+
+- Remove deprecated `purl_to_lookups` and `without_empty_values` import compatibility
+  from `packageurl.contrib.django.models`.
+  Replace those functions import using `packageurl.contrib.django.utils`.
+- Add download purl2url support for bitbucket and gitlab.
+
+0.11.1 (2022-03-24)
+-------------------
+
+- Add support for the golang type in `purl2url.get_repo_url()` #107
+
 0.11.0rc1 (2022-12-29)
--------------------------
+----------------------
 
 - Apply typing
 - Add support for Python 3.11
@@ -11,14 +35,13 @@ Changelog
 
 
 0.10.5rc1 (2022-12-28)
-------------------------
+----------------------
 
-- Fixed `PackageURL.from_string` to properly handle npm purls
-  with namespace.
+- Fixed `PackageURL.from_string` to properly handle npm purls with namespace.
 
 
 0.10.4 (2022-10-17)
---------------------
+-------------------
 
 - Refactor the purl2url functions and utilities #42